set initial index to snap position, shared screen for profile and dashboard beat expansion
