+++ app/[drop].tsx
@@ -1,0 +1,2 @@
+import drop from "@/components/drop";
+export default drop;
