--- app/profile/[drop].tsx
@@ -1,2 +1,0 @@
-import drop from "@/components/profile/drop";
-export default drop;
