--- app/profile/_layout.tsx
@@ -1,3 +1,0 @@
-import { Slot } from "expo-router";
-
-export default Slot;
